refactor path integration utils
